FaceDetection/requirements.txt
streamlit
deepface
tf-keras
opencv-python-headless
numpy
pillow
pandas